consumer/requirements.txt
python-dotenv==1.0.0
kafka-python==2.0.2
pymysql==1.1.0
cryptography==41.0.7